[!!!][BUGFIX] Drag and Drop bugfix
This bugfix consists of several precautions:

* A new colPos of -42 has been introduced for "Content Container";
* New colPos replaces MakeQueryArray conditions for cleaner approach.
* New ColumnPositionListItemsProcessor extends EXT:gridelement one.
* EXT:gridelements has been set as a dependency to avoid a conflict.

In order to avoid a conflict with tt_content.colPos itemsProcFunc
which can only be defined once and overrides the previous selection,
EXT:gridelements must now be loaded BEFORE EXT:flux or a "INVALID
VALUE (-42)" label will appear in the "column" selector box. Although
this is only cosmetic it is one reason for marking this as a breaking
change. This does not affect new installs - only ones where Flux was
originally installed before EXT:gridelements.

The other reason for marking as breaking is that after this patch
you may experience non-cosmetic problems if you previously had
content elements that previously belonged to a Flux content carrying
element (or FED FCE w/ nested content elements) but was not properly
deleted or moved for some reason, these will not be visible again
and must either be hidden or deleted. FE-templates using FED's
page.renderContent ViewHelper are not affected but others rendering
methods will probably display the previous hidden content in FE.

Fixes: #34769
Fixes: #34718
Fixes: #33338
Change-Id: Ia5e4519d651a28174c36829215499a655bcd80fe