[TASK] Migration to FieldProvider pattern
Changes the internal logic that is rendering FlexForm structures
from the previous XML-to-TCA to a direct TCA approach.

Briefly described this change...

* Completely removes the intermediate XML step and all XML templates
* Removes all partial templates previously used to render field XML
* Introduces FieldProviders which function similarly to ConfigurationProviders
* Improves performance slightly by skipping two expensive rendering steps
* Streamlined a classes and viewhelpers

Also contained are a few additional ViewHelpers and some basic bugfixes.
Apologies for the combined commit; after working on this for three months
it was too complicated to split this into smaller pieces.

Note: fixes #37190 but cannot guarantee other extensions will not
re-specify tt_content.colPos as something other than signed smallint(6).

Change-Id: Ic9812b2af83a6a142b48a886617c3458c45d4dad
Fixes: #39269
Fixes: #37190
Fixes: #36907
Fixes: #34543
Fixes: #35530
Fixes: #38647
Fixes: #34717
Fixes: #34926
Fixes: #38434