[FEATURE] Add deferred data processing hook-style methods to Providers
This change introduces new methods on ConfigurationProviderInterface
implementing classes which will trigger on key operations. When a
ConfigurationProvider has been registered to a table (optionally
also a single fieldname that must be included in order to trigger)
the according methods will be triggered from TCEmain when manipulating
record data.

Note: this is the basis for refactoring the Flux FCE processing
logic out of the TCEmain class, as well as the basis for a new
feature in FED which will filter out FlexForm data that is identical
to the parent's data - in order to preserve inheritance of data
for page configurations even if a child page has been saved with
a configuration that is tied to just that page but is identical
to the parent page's configuration.

Although this commit is set to fix #33732 a new commit in FED
will be made using the same issue no. as reference. Until that time
the referenced issue is not entirely done.

Change-Id: Ice222cc7f8bf87e2b03b1b31e8d212b455672eee
Fixes: #33732