[TASK] Move definition of default RTE configuration to TypoScript
This change adds a new TS paramter at location:

* plugin.tx_flux.settings.flexform.rteDefaults

This location can be overridden to change the default RTE configuration.
Fallback behavior is to act as before with one exception: if for any
reason you previously defined "defaultExtras" on flux:flexform.field.text
and depend on fallback to the Flux default if that value is an empty string,
you must either use this new TS option or add manual fallbacks to your
templates so a proper RTE configuration will always be present - if not,
the TextStructureProvider will not receive a NULL value as "defaultExtras"
and will thus not trigger fetching of the TS configuration.

Change-Id: Iaf330fa55ea4d8449ceb64b6c42c2e0824becd53
Fixes: #33582