[BUGFIX] Avoid fallback StructureProviders if TemplaVoila is loaded
This change adds a condition not to use fallback data structures if
TemplaVoila is loaded. Since TemplaVoila uses the same hooks as Flux
does Flux must consider the possibility that TemplaVoila is loaded
and if it is, cut away some of the developer friendly fallbacks in
order to get compatibility.

Change-Id: I095ddc685d75e2614a2abb194ce0b3efaef1e162
Fixes: #39522