[TASK] Refactor TceMain handling into ConfigurationProvider
Developer hint: ConfigurationProviders are now complete handlers
for records and record fields. You can use a ConfigurationProvider
to transform record data saved in any table, even ones belonging
to core extensions. For example, one Provider can be triggered
whenever a tt_news record is being saved, moved, deleted or sorted,
another could trigger when the same is done to a tt_news category.

This refactor adds additional methods on the ConfigurationProvider
base class and interface, allowing methods to be executed by the
TceMain hook based on which table is being saved to / operated on.

Also changed is that tx_flux_column and tx_flux_parent are now
editable fields; the select box containing content areas only
contain areas that are actually defined in the content element.

Also added is a ContentConfigurationProvider which contains the
refactored methods from the TceMain hooks class. The new Provider
is similar to the ContentConfigurationProvider in FED but where
the FED Provider defines how (template-wise) to render a acontent
element, this Provider defines what happens when saving or when
re-arranging ccontent elements.

Using the methods on the ConfigurationProvider corresponds to using
a full TceMain hook class. Each method is executed when a particular
action is taken, allowing the implementer to change record values
and perform additional processing.

It is possible to register multiple ConfigurationProviders for one
table and even one fieldname (and even with the same priority).

Fixes: #39974
Fixes: #39973
Fixes: #39633
Change-Id: I8910085f7aeb8133d1ed30587b8358f0a04b66ff