[FEATURE] (SwitchableFlexForm)ControllerActionsViewHelper
ControllerActions ViewHelper

Renders a FlexForm select field with options fetched from
requested extensionName/pluginName and other settings.

There are two basic ways of adding selection options:

- You can use the "extensionName" and "pluginName" to render all
  possible actions from an Extbase plugin that you've defined. It
  doesn't have to be your own plugin - if for example you are
  rendering actions from EXT:news or another through your own plugin.
- Or you can use the "actions" argument which is an array:
  {ControllerName: 'action1,action2,action3', OtherControllerName: 'action1'}

And there are a few ways to limit the options that are displayed:

- You can use "excludeActions" to specify an array in the same
  syntax used by the "actions" argument, these are then excluded.
- You can specifiy the "controllerName" argument in which case
  only actions from that Controller are displayed.

And there are a couple of ways to define/resolve labels for actions:
- You can add an LLL label in your locallang_db file:
  lowercasepluginnanem.lowercasecontrollername.actionfunctionname
  example index: myext.articlecontroller.show
- You can do nothing, in which case the very first line of
  the PHP doc-comment of each action method is used. This value can
  even be an LLL:file reference (in case you don't want to use the
  pattern above - but beware this is somewhat expensive processing)
- Or you can do nothing at all, not even add a doc comment, in which
  case the Controller->action syntax is used instead.

Marking actions that have required arguments (which cause errors if
rendered on a page that is accessible through a traditional menu) is
possible but is deactivated for LLL labels; if you use LLL labels
and your action requires an argument, be user friendly and note so
in the LLL label or docs as applies.

Lastly, you can set a custom name for the field in which case the
value does not trigger the Extbase SwitchableControllerActions feature
but instead works as any other Flux FlexForm field would.

To use the field just place it in your Flux form (but in almost all
cases leave out the "name" argument which is required on all other
field types at the time of writing this). Where the field is placed
is not important; the order and the sheet location don't matter.

Change-Id: I9ac66d5f4dee2ace3059b5dbec1f0a2f0e11c730
Fixes: #40246