[BUGFIX] Ignore Provider field names on inconsistent TYPO3 versions
This change patches various places in the code to handle TYPO3 versions
which do not consistently pass along the field name that is being
processed. In these cases the field name must be ignored - but this
does however have the unfortunate side effect that only one Provider
can be registered per table and each Provider must be able to determine
wether or not to process the data structure. Which is what is currently
being done.