[TASK] Added legacy template for Grid Widget
This template is the previous template, which can no longer be made compatible with the new native drag and drop feature. The choice was between using two sections to contain the new/legacy renderings or directly rendering a faked action, "legacy", which renders a completely separate template but with all the same variables. This latter solution was chosen.