[BUGFIX] Set template path instead of setting action as legacy support, Grid Widget Controller
For some extremely odd reason, setting the action in order to render a different template has peculiar effect of allowing the Grid to function with any typical content element - but then completely self-destruct whenever another Flux-enabled content element is inserted in the Grid.

Setting the template path and filename (but then also preventing override of the Grid template on legacy versions) avoids this and fully restores support for nested (and nested-nested-nested-etc.) Flux-enabled content elements.