[BUGFIX] After saving FCE, nested content had invalid tx_flux_column
This problem would occur on 6.0 because it uses a section in the returnUrl; something that normally would be used by Flux to indicate where to create the content.

However, on 6.0 this returnUrl processing is no longer necessary which means that returning a NULL value for area:parent is safe. On 4.5 -> 4.7 behaviour is unaffected (added though, by this, is increased URL error tolerance).

Fixes: #14
Fixes: #15