[FEATURE] Support NULL arrays when resolving ConfigurationProviders
Allowing NULL instead of an array causes the parameters to be loosened slightly, returning any ConfigurationProviders for the table in question.

It now becomes up to the ConfigurationProvider to assert wether the array actually has usable values.

Currently there is only a single case where a NULL value is passed: when executing the `clearCacheCommand` method on ConfigurationProviders.