[FEATURE] Flux variable sliding based on parentFieldName
This first phase allows simple, unconditional inheritance with merging of variables based on the inheritance tree. Also added is processing which strips off values which are unchanged in the local element, saving only the diff of inherited+local settings.

This means that in order to be fully restored, variables must also be retrieved by the ConfigurationProvider which then takes care of re-parsing the inheritance tree, merging variables to construct the proper full set of relevant variables.