[TASK] Remove type requirement from hook methods in TCEmain
This should allow 6.0 and 4.x to use the hook methods also on systems where the nasty class mapping bug is in effect. In this case, the solution appears to work - since the TCEmain hook class is not hooked up to an interface, it is possible to avoid a type requirement.