[FEATURE] Support "paste" and "paste as reference" through clickable icons
These icons become visible when there is content in the TYPO3 clip board. When copying elements, they can be pasted as reference which creates a new element type "shortcut" pointing to the actual content element. When moving content, references cannot be created and thus the "paste as reference" icon is not displayed.