[FEATURE] Debug Service
This adds a new DebugService (which is also prepared for, but not yet ready to, do advanced debugging on ConfigurationProviders and ExposedViews) and implements that Service in every place where the debug flag was previously used to suppress warnings.

This **significantly** increases the usability of Flux based templates in a development context: the error messages are displayed as flash messages and even errors reported from AJAX-triggered hooks can be viewed in the backend.