[TASK] Simplify field structure compilation process
This change applies a huge cleanup in field structure (TCEforms) compilation when using Flux field ViewHelpers. Before, a collection of special StructureProviders were used - now, this all happens inside the ViewHelper.

There is still a lot which can be cleaned and simplified even further - for example, the configurations being built definitely contain redundant variables. But this initial simplification cuts away a lot of redundant source code and makes the logic easier to read.