[BUGFIX] Using Flux with other v4-only hook subscribers
This fixes an issue with expected class names in hook subscribers which either don't want to or can't remove a forced type indication for the "caller" parameter.

Before, Flux extended the new class names and since the new class names don't extend the old ones (it's vice versa), hook subscribers which have a hard expectation of an old class, would fail because of receiving a new class.

After, Flux now subclasses the old class which subclasses the new class that Flux used to subclass - result is same behaviour as now, one additional level in the inheritance tree and compatibility with v4-only hook subscribers.

Fixes: #125