[FEATURE] Thin API to add global, static TypoScript includes
These inclusions do not show up in the static TypoScript template selector - and thus should only be used for truly global TypoScript which is 1) not susceptible to inclusion order and 2) never needs to or is allowed to be removed by an integrator.

Integrators can still change the TypoScript values included this way - but cannot remove the entire inclusion.