[DOC] Document new strategy to use "mergeValues" behavior
Attribute "mergeValues" on flux:flexform is deprecated and has no counterpart in the FormComponent. The description has been changed to inform of a special record processing feature in flux: prefixing field names with the table name causes automatic overriding, e.g. "tt_content.header" as name of Flux field results in "header" in the record being overwritten with the Flux field value.