[TASK] Move flexform-to-record value merging to bottom class
The feature is now completely generic: merging values can happen in any table. For example, a Provider for table/column "fe_users.username" will cause the record's column(s) to be overwritten with flexform data if flexform fields are prefixed with "fe_users." - which means a flux field name of "fe_users.username" will write the value of the field to the "username" column in the record when the data is saved.