[TASK] Move LLL handling to base Form Component and Interface
This enables the use of LLL-affecting properties on every Form Component type and ensures that if set on the parent container BEFORE a child is created from this container, these values will be transferred to the child. This makes it possible to change the location of the LLL file on-the-fly, for example using one LLL file per sheet (which may help if your form is huge).