[TASK] Refactor and extend type support in transformation logic
This fixes a few issues with namespaced Domain model class names and repositories - and enables FluxService to test for presence of a method called "findByIdentifiers" on Repository instances which will be used to fetch objects by an array of identifiers (exploded from field value). Implement this method when you need to control exactly how your objects get fetched; if the method is not there the usual approach (create Query, match manually, execute) is selected.