[FEATURE] Let Pipes return Form Fields for own properties
This allows any Flux Provider+Form which uses Pipes, to load a Pipe instance and get a set of basic Form Field instances which can be attached to a Sheet, Container, Object etc. to insert fields which when saved, result in an array that can be used as argument for the `loadSettings` method on all Pipes, to configure that instance with all settings that were entered in the Flux Form.

Required by EXT:fromage.