[FEATURE] Custom ConfigurationManager with multiple PID resolves
Adds a new ConfigurationManager which loads a custom BackendConfigurationManager which contains multiple methods of resolving a page UID based on different URL arguments. Covers most cases in the TYPO3 backend where a PID does exist in some argument, but is not located in the `$_GET['id']` argument which is the only one the basic Extbase ConfigurationManager respects.

PID can be resolved from:

* GET.id
* POST.id
* GET.edit.$table.$pid (when *creating* records, $pid is expected PID)
* GET.edit.$table.$id (when *editing* records, $id is UID of record: load record from DB and read PID of record)
* GET.edit.$table.$id (when $table is `pages` then $id is PID, both when creating and editing pages)
* Currently active ContentObject record (very rarely filled in BE, included for completeness)
* TypoScript record's PID *if* there is only one active root TS record (which is how the current Extbase fallback works)

The result: failsafe PID resolving not just in Flux but every other extension, when Flux is installed.