[FEATURE] White/blacklisting allowed CTypes in Column and Area
Limits possible selections in the new content element wizard based on parameters defined in the Column or Area (or both) into which content is being inserted.

## Usage examples

In a content/plugin template:

```xml
<flux:grid.row>
	<flux:grid.column>
		<flux:form.content name="content">
			<flux:form.variable name="allowedContentTypes" value="text,textpic" />
		</flux:form.content>
	</flux:grid.column>
</flux:grid.row>
```

In a page template:

```xml
<flux:grid.row>
	<flux:grid.column colPos="1">
		<flux:form.variable name="allowedContentTypes" value="text,textpic" />
	</flux:grid.column>
</flux:grid.row>
```

Alternative syntax:

```xml
<flux:grid.row>
	<flux:grid.column colPos="1" variables="{allowedContentTypes: 'text,textpic'}" />
</flux:grid.row>
```

Using PHP (Flux Form Components):

```php
$grid = \FluidTYPO3\Flux\Grid::create();
$row = $grid->createContainer('Row');
$column = $row->createContainer('Column');
$column->setColumnPosition(1)
	->setVariable('allowedContentTypes', 'text,image');
```

To blacklist, use `deniedContentTypes` instead of `allowedContentTypes` as variable name. Both can be used together: first, the whitelist is applied if it has any types contained, and second, the blacklist is applied to remove elements from the list of allowed types (which is either the full list or the previously whitelist-filtered version, if a whitelist was specified).

Fixes: #197