[BUGFIX] Nested records are not copied right
If you copy an Row Element with nested Content Elements the information of the the column of the nested Elements is lost and the nested Elements are displayed beside the Row Element and not inside any more.

This happens because tx_flux_column of nested records is emptied as well. During my fix i encountered a few more problems with copying and moving of nested Elements and i think i found a solution for them all.

Please check this commit! =)