[TASK] Avoid DataMapFactory in TableConfigurationProcessor
This avoids the instanciation of a cache which, at the point of this execution, may or may not be alive and filled with a proper data map. The result being that depending on execution sequence and utilized hooks, Extbase data mapping based on table columns and TCA was made impossible (always empty column map).