[BUGFIX] Improve integration of workspaces child content
This patch fixes a few issues with respecting the record selection and display conditioning done in PageLayoutView, as implemented in Flux's content area renderer which selects the content elements to be displayed.

Most importantly this fixes the incorrect display of move and delete placeholders, correctly filtering deleted records out and using the placeholder's position values when placeholder is a move placeholder.

References: #627