[TASK] Use core BE configmanager getPageUid as fallback
This ensures complete compatibility compatibility with CLI commands which fall back to loading root page UID from the root TS record. Before, the behavior of this resolving was a bit too strict but perfect for BE usage - after, the final, custom TS record resolving is replaced with the (only) method the standard BE configmanager normally uses.

Fixes: #668