[BUGFIX] Field.TextViewHelper should respect global RTE Settings by default
If you enable the RTE and set defaultExtras to an empty string the resulting Behavior of the RTE
is to show everything. This commit changes sets the defaultExtras to "richtext[]" if the option
enableRichText is set to true, but defaultExtras contains an empty string. "richtext[]" instructs
the Form to use the globally defined RTE Settings.