[TASK] Rename "extensionName" on ControllerActions Field ViewHelper
THIS IS A BREAKING CHANGE.

You must adapt templates which use the `flux:field.controllerActions` ViewHelper with the `extensionName` argument and change the `extensionName` argument usage to `controllerExtensionName`.

No change is required if you do not use the `extensionName` argument; automatic detection still works as before. The original argument `extensionName` is still possible to use, but now has another function (setting the extension context for LLL generation).