[TASK] Avoid re-rendering ExposedTemplateView sections
Lets getStoredVariable() check for an internal registry of stored instances of ViewHelperVariableContainer and retrieve the variable from it if it exists.

Avoids expensive re-rendering if the same View instance is used to retrieve the same variable multiple times, or used to retrieve multiple variables from the same section (e.g. form and grid).