[TASK] Make automatic icon resolving more resilient
This change modifies the way template paths are analysed to determine which icon file locations to check for the presence of icons. Before, paths were incorrectly trimmed if path contained dots. Now, only path separators are taken into consideration and the file path is treated from the end of the string only.