[FEATURE] Implement ViewContext and TemplatePaths objects
These objects are used as containers to define a context (extension key, variables, controller name, format etc) for rendering a view, and a collection of template paths (root paths for partials, layouts and templates) which are automatically parsed from all supported input structures to a single structure; the new templateRootPaths etc. (plural paths) approach.

Implementation can still be cleaned further but the new classes themselves are in a proper, working and fully test covered state.