[FEATURE] Add translation capability to Flux forms
### WARNING! DO NOT SWITCH THIS VALUE FOR EXISTING CONTENT

Changing the value of this flag causes TYPO3 to look for the value of each field in a different location and should never be done for templates that are associated with content that is currently in use. Changing the behavior will NULL all values.

The default is set to the current Flux behavior which is to disable translation completely.

Usage examples:

```
<flux:form id="myform">
    <flux:form.option name="translation" value="inherit" />
</flux:form>

<!-- same as: -->

<flux:form id="myform" options="{translation: 'inherit'}">
</flux:form>
```

Possible option values are:

* `disabled` (default)
* `separate` which adds a complete copy of the entire form for each language and expects a full translation to be entered
* `inherit` which adds copies of each field for each language and supports fetching the default value if no translated value is provided