[BUGFIX] Adjust parent UID during localisation of record
Before this patch, localising a single record that was a child of another record would cause the localised record to contain the original parent's UID in the `tx_flux_parent` field, causing incorrect display in the page module. Fixing the value of this field during initialisation of the localised record puts the record in the right parent.