[BUGFIX] Fix wrong plugin type matching
If a user inserts and saves a "General Plugin" as CE without specifying the list_type the
preview in Backend breaks (happens easily with the quick ce button from gridelements).
This is caused by a falsey matching the correct plugin type if it's empty. As the check for
plugin type is only needed if the row actually contains a plugin we can remove the case
of an empty list_type.