[BUGFIX] Fix for not showing hidden content elements in grids anymore
TYPO3 seems to have changed the storage of "Show hidden content elements" recently, which causes fluidcontent to not get the flag to
show of hide hidden content elements inside grids, etc.
TYPO3 now defaults to showing hidden content elements if the flag is
not set. To adapt to this this commit adds the same behavior to default
to shoing hidden content elements. as soon as you switch the "Show hidden content elements" to off and on again it works, because then the flag
is present in the session.