[TASK] Adjust to unannounced core behavior re: renderMode
The core now renders an *error* if the `renderMode` attribute of a `select` TCA type is set to not-NULL, where before, it would silently allow the perfectly valid configuration array to be used nonetheless.

In addition, the error message displayed says "Unsupported type: select" which is obviously incorrect. The right error message would have been "Unsupported select renderMode: default".