[BUGFIX] Fixed incorrect cleaning of FlexForm XML
During FlexForm cleaning, field nodes are iterated over and specially
flagged fields are removed. This removal happens during iteration and
the change in the array causes the next element to be skipped.

This is bug is visible within EXT:fluidpages when saving a page with
no configuration changes. Values that were not specifically set on the
configuration form are saved in the FlexForm XML.

The solution is to gather fieldNodes to be removed, but remove them in a
separate loop.