[BUGFIX] Prevent Form component from becoming child of self
Earlier, Flux ViewHelpers would resolve components *and* associate them with the parent in the same method. Now, this is decoupled in order to provide a shared `renderStatic` method - but in doing so, an edge case slipped in where any ViewHelper which did not create fresh but rather resolve existing Form components (which is the case of Sheet when used with a name of an already existing sheet), could now possibly become a child of itself, causing infinite recursion when attempting to use `get()` with recursive mode on such a component.

Close: #1041