[BUGFIX] Patch TCA when applied in FlexForm context
This change patches the data source array generated by Flux form components when the data source array is consumed by FlexForms in TYPO3, by adding an additional "TCEforms" array dimension around the TCA configuration.

The behavior is put behind a compatibility registry flag to allow it to be made version dependent in the future. FormEngine will likely change on this point - and/or Flux will provide an alternative to the traditional FlexForms data source approach. This is why the change is implemented as a patching operation in this context only - for all other TCA contexts the configuration generated by Flux components is valid without the "TCEforms" dimension.

Close: #1047