[BUGFIX] Patch for new FormEngine and IRRE editing
This patch updates a few tests for compatibility and works around a regression in the TYPO3 core which caused IRRE AJAX requests to fail when the data source involved any type of hook to manipulate the data source and that hook requires the complete database record. In addition handles newly createed records by reusing a previous data source that gets cached with an identity based on the TCA configuration of the parent and field.

It may not be pretty and it may have some rough corners when dealing with data structures that (for example) are supposed to be dynamic also when creating new records - but it is a big improvement over the current "Server reported an error (500)" javascript popup that gets displayed.

Also works around the issue for other similar implementations when Flux is installed - e.g. does not only target Flux data structures.

The plan for patching this in TYPO3 itself still exists but involves a significant amount of rewriting javascript - the intention is to include the entire data structure in the AJAX request, which implies somehow storing it in the DOM and have the javascript extract it before making the IRRE request. In the meantime we can rely on this imperfect workaround.

References: https://github.com/FluidTYPO3/fluidcontent_core/issues/207
References: https://github.com/FluidTYPO3/flux/issues/1013
References: https://github.com/FluidTYPO3/flux/issues/1033
References: https://github.com/FluidTYPO3/flux/issues/1055
References: https://github.com/FluidTYPO3/flux/issues/1069
Replaces: https://github.com/FluidTYPO3/fluidcontent/pull/321