[FEATURE] Register extension key for any controller/feature
Allows `\FluidTYPO3\Flux\Core::registerProviderExtensionKey` to be called with a single argument (extension key - Vendor.ExtensionName format) which means the extension gets registered as a provider for *every* possible feature.

Whether or not an extension registered this way then adds resources for a given feature depends solely on the presence of folders/files/templates as defined in FluidTYPO3 conventions.