[BUGFIX] Catch and silence StopActionException in controllers
Since we call controllers without going through the request dispatcher this exception is not caught for us. And we can't use the dispatcher to dispatch our request because that will cause controllers to be resolved which we've already done.

Solution is to do as dispatcher does: catch and ignore this particular exception type.

Close: #1039