[BUGFIX] Translation: Only set container if original does
When translating content elements, we now only register the
translated content record to a container if it the translation
original has been registered (tx_flux_parent > 0).

Before, if no container was used, a search occured over all
content elements with t3_origuid = 0 (meaning: all non-copied
and untranslated records) and the newly created translation
would be registered to that container (even if it was no
container FCE at all or on a completely different page).

This patch also eliminates the need for the previously used
foreach loop by querying database directly for the wanted
sys_language_uid.

To be more clear about what actually happens,
$translatedParent(s) has been renamed to $copiedParent(s):
translated containers are currently being found by looking
for *any* copy of the original language container, not actual
translations which may cause more bugs and should be fixed...

This commit should resolve the root cause of #1125 but will not
clean up the database corruption it already produced. If pulled,
it may supersede part of the pending commits from #976.