[FEATURE] Update script shows inaccessible content due to #1125
The update script now checks for any content in COLPOS_FLUXCONTENT
without a flux parent or column reference, i.e. content previously
"lost" due to #1125. All found (and not deleted) content is being
listed and can be edited. Instructions on how to fix the issue are
included with that listing.

Note that there may be other cases where content becomes inaccessible.
It may be a good idea to check for those as well and include them in
that listing with amended instructions.