[BUGFIX] Require older test tools in legacy branch
The legacy branch currently requires the newest version
of the FluidTYPO3 Development Tools.

This version requires TYPO3 7, the legacy branch however
is based on TYPO3 6. Therefore the runtests command
is not executable in the legacy branch.

To be able as a contributor to send usefull PRs to the
legacy branch the test should still be executable.

Therfore an older version of the test tool should
be required in the legacy branch.