[BUGFIX] Reproduce solution in http://forge.typo3.org/issues/78353 on 8.4.x
The patch on http://forge.typo3.org/issues/78353 is unfortunately
not present on the 8.4.x branches of TYPO3. Luckily, a fairly
simple "hack" can be used, targeting this specific version to add
an additional AND suffix to the custom condition we use when
pulling content records in a given Flux content area.