[BUGFIX] Fix DataViewHelper when injectors don't execute
In fully compiled templates the injection methods will not
execute since a VH class instance is never created. The
classes used are now wrapped in a getter method to create
instances as needed.

NB: inject annotation removed, injection method preserved
in order to allow replacing instances from the outside in a
way that still works with uncompiled rendering.

Close: #1228